Reworked a cleaner Makefile, and corrected IDL wrappers.
The new Makefile uses configurations in the `compile` directory, of the
format `$(OS)-$(ENV).make`, which defines the compilers and options. We
should add more configuration files for each target architecture.

The IDL wrapper has been externalized in a single C interface file.
Tested on GNU and Intel compilers on Linux.